+++ README.md
@@ -17,15 +17,15 @@ whether communication is currently active.
 For example, to display status of the Rx button for LMT_GND, as shown in the image above, configure the
 station status button like this:
 
-![Screenshot of a station status button configuration, with callsign set to LMT_GND, RX selected, and three custom images specified for the three states](docs/images/station-status.png)
+![Screenshot of a station status button configuration, with callsign set to LMT_GND, RX selected, and three custom images specified for the three states](docs/images/stationStatus.png)
 
-| Setting | Description |
-| - | - |
-| Callsign | The callsign for the station you want to display status for. Required. |
-| Listen to | What status to display on the button, either Rx, Tx, or XC. Required. |
-| Not listening | The image to display when the station is not currently active. Optional, defaults to black. |
-| Listening | The image to display when the station is active. Optional, defaults to green. |
-| Active comms | The image to display when a transmission is actively taking place. Optional, defaults to orange. |
+| Setting       | Description                                                                                      |
+| ------------- | ------------------------------------------------------------------------------------------------ |
+| Callsign      | The callsign for the station you want to display status for. Required.                           |
+| Listen to     | What status to display on the button, either Rx, Tx, or XC. Required.                            |
+| Not listening | The image to display when the station is not currently active. Optional, defaults to black.      |
+| Listening     | The image to display when the station is active. Optional, defaults to green.                    |
+| Active comms  | The image to display when a transmission is actively taking place. Optional, defaults to orange. |
 
 ## Configuring a TrackAudio status button
