+++ com.neil-enns.trackaudio.sdPlugin/images/actions/trackAudioStatus/connected.png
