+++ com.neil-enns.trackaudio.sdPlugin/images/actions/trackAudioStatus/notConnected.png
